feat: 레이싱 자동차 객체
